requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
npm==0.1.1
numpy==1.26.4
openai==1.25.0
optional-django==0.1.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
prompt_toolkit==3.0.46
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.20.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers==4.41.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
wsproto==1.2.0
